사용자 즐겨찾기 목록

+++ src/main/java/com/mysite/sbbshopapi/mapper/FavoriteMapper.java
@@ -9,5 +9,5 @@ import com.mysite.sbbshopapi.dto.Favorite;
 @Mapper
 public interface FavoriteMapper {
     void insertFavorite(Favorite favorite);
-    List<Favorite> selectFavoritesByUserId(int userId);
+    List<Favorite> selectFavoritesByUserId(Long userId);
 }
